Step2_1_Pinecone_RAG/requirements.txt
pinecone-client
sentence-transformers
pandas